Separated limit, usage, and requested for CPU and memory
